+++ CHANGELOG.md
@@ -7,10 +7,8 @@ in 0.1 minor versions.
 
 To get the diff for a specific change, go to https://github.com/superdesk/web-publisher/commit/XXX where XXX is the change hash
 
-* 0.2.1
- * feature [#211] Create Automatic list widget 
-
 * 0.2.0
+ * feature [#211] Create Automatic list widget
  * feature [#209] Add option to ignore context meta in loaders 
  * feature [#209] Add rendition loader (gimme rendition) from article media
  * feature [#203] Added option to access theme assets with simple /public/{fileName} links